updated the underlying library

+++ src/Timer.cpp
@@ -3,7 +3,7 @@
 
 Timer::Timer() { this->interval_end_ms = 0L; }
 
-Timer::Timer(int ms) { countdown_ms(ms); }
+Timer::Timer(int ms) { countdown_ms((unsigned long)ms); }
 
 bool Timer::expired() {
   return (interval_end_ms > 0L) && (millis() >= interval_end_ms);
@@ -15,4 +15,4 @@ void Timer::countdown(int seconds) {
   countdown_ms((unsigned long)seconds * 1000L);
 }
 
-int Timer::left_ms() { return interval_end_ms - millis(); }
+int Timer::left_ms() { return (int)(interval_end_ms - millis()); }
